R Code for makeCacheMatrix and cacheSolve - Assignment #2
This set of R code functions is designed to invert a solvable matrix and
save the inverse to cache.  makeCacheMatrix(x) creates a special matrix
with a series of sub-functions.

cacheSolve(x) will check whether the matrix has been previously inverted
and stored.  If so, it will retrieve and return the cached value saving
on processing.  If not it will calculate the inverse, displaying and
caching the results.  This is the version 1.0 of the code for review.